TASK improve accessibility of cards and tableview
To enable keyboard users to show server details the links are added as
HTML `a` Tags not only for <noscript> variant.
This also improves accessibility as screen reader users usually are not
able to perform click events on div elements. The do not even know there
is a click event handler. Click event handlers only work on buttons by
default as browsers forward the keyboard activity on buttons to the
onclick handler.