Update Octopush.php
With the low cost default value, SMS are sometimes received many hours after. For a monitoring software, this has no sense, we need to be sure to receive the alert by SMS instantly to repair the problem as quick as possible. Otherwise, SMS is useless, email is free and good enough. Moreover, the price difference between low cost and premium is small.