added support for connectivity_sanity_test
If the user has entered an IP address for the connectivity_sanity_test field on the server edit page, that IP address will be pinged before and after that server gets checked.  If either of those pings fails, the server uptime results will not be logged because it is assumed that the phpservermon system is experiencing a connectivity problem.  If both sanity test pings succeed, the server uptime results will be logged as normal.