Remove onclick from non-link table cells
This removes the onclick attribute from the table cells on the index page.  The first cell of each row already has an "a" link which takes the user to the same place and is more intuitive since hovering over that "a" link changes the cursor to a "pointer."  The other cells still showed a "text" cursor which implied the elements were not clickable, though they were.  Removing the onclick from those other cells also allows the user to select their text for copy/paste purposes without getting redirected to another page unexpectedly.